Made recommended changes
Notes on Low Level Code 

Good organization with Vehicle, Brake, ThrottleController etc. classes. 

The only thing in the main loop is vehicle.update, which checks the brakes, computes the speed and the steering angle.  It does not set the throttle, so there is no way to make the trike go. It does not set the steering angle or apply the brakes.  How is the vehicle controlled?  Where is communication with the Remote Control or High Level? 

I added the existing communication code from the elcano/elcano repo but this is the section we will be reworking for can 

There are two reasons to avoid doubles. I would prefer to see doubles eliminated from the code, except where you need to interface to a library function. 

On most Arduinos (Due excepted) double and float are both 32 bits.  Be honest that you are just using 32 bits. 

The Atmel AVR does not have hardware floating point. It can do arithmetic with a long int much more quickly than it can manipulate a float. Almost all calculations can be done with ints (paying attention to underflow and overflow). 

I replaced most of the doubles, the PID requires the type double for some of it’s inputs. We can mess with that lib b 

Not having PID is OK for an interim solution, but not acceptable for the final product. There must be speed control. It does not have to be PID, but there must be a control method that sets the throttle to maintain a desired speed. 

The PID controlls have not been tested but they can be enabled by using the myTrike.move instead of myTrike.noPID 

Lines 38 and 39 of Vehicle.cpp: 

if (dSpeed < (throttle.getSpeedInput_mmPs() + 10))  // should it be >   ? 

brake.Stop(); 

This is weird code that existed previously, and I would like to ask where the condition came from as dSpeed is the desired speed in cmPs and getSpeedInput is the last measured speed in mmPs 

In Elcano_C2_LowLevel, WheelRev is an ISR.  It does a call to Vehicle.tick which calls thottleControl.tick. These calls add to the time in the ISR. The ticks in the Vehicle and ThrottleController classes are not obviously interrupt routines.  I would prefer to have WheelRev do its business directly without the extra calls.  tickTime_ms needs to be volatile. 

I have changed tickTime_ms to be volatile. I am not sure how to eliminate the additional calls. I can eliminate the one inside throttle by making the tickTime_ms array public but making the entire throttle object public within vehicle feels like a bad design choice.  

 

These two settings files are obsolete and have been removed. 

Trike_settings.h: 

There are multiple units being used for the steer and brake actuator. 

The _OUT numbers are 0 to 255, intended for an Arduino PWM signal to control the actuators.  

The _COUNT numbers are in ms. 

The _US should be microseconds and are not compatible with _OUT 

Lines 13-15 for the brake actuators are obsolete. 

Lines 5-32 should be eliminated. 

Lines 51-53 should be moved to pin_settings.h 

Pin_settings.h: 

The analog output is on lines 6,15,35-37 and should be grouped together. 

Eliminate lines 9-14. 

The IRPT_ numbers belong to the RC receiver; eliminate lines 16-33.

+++ Elcano_C2_LowLevel/SteeringController.h
@@ -9,14 +9,14 @@ class SteeringController{
 	double steerAngleUS;
 	double PIDSteeringOutput_us;
 	double desiredTurn_us;
-	int currentSteeringUS =0;
-	void turnOn(double input);
-	void SteeringPID(int input);
+	int32_t currentSteeringUS =0;
+	void turnOn(int32_t input);
+	void SteeringPID(int32_t input);
 public:
 	SteeringController();
 	~SteeringController();
-	void initialize(double input);
-	void engageSteering(double input);
-	void updateAngle(double input) { steerAngleUS = input; };
-	void setDesiredTurn(int input) { SteeringPID(input); };
+	void initialize(int32_t input);
+	void engageSteering(int32_t input);
+	void updateAngle(int32_t input) { steerAngleUS = input; };
+	void setDesiredTurn(int32_t input) { SteeringPID(input); };
 };
